move slash command handling into respective files
