+++ src/events/interactionCreate.ts
@@ -6,7 +6,8 @@ import {
   MessageSelectOptionData,
 } from 'discord.js';
 import TurndownService from 'turndown';
-import { getFaq } from '../slashCommands/faq';
+import { getFaq, handleFaqInteraction } from '../slashCommands/faq';
+import { handleTicketInteraction } from '../slashCommands/ticket';
 import { event } from '../types/event';
 import Client from '../util/Client';
 
@@ -14,8 +15,6 @@ const interactionCreate: event<'interactionCreate'> = async (
   client: Client,
   i: Interaction
 ) => {
-  const turndownService = new TurndownService();
-
   if (!i.isCommand()) return;
 
   if (i.isApplicationCommand()) {
@@ -23,73 +22,12 @@ const interactionCreate: event<'interactionCreate'> = async (
 
     switch (i.commandName) {
       case 'faq': {
-        switch (i.options.getSubcommand()) {
-          case 'list': {
-            const faq = await getFaq();
+        await handleFaqInteraction(client, i);
+        break;
+      }
 
-            const embed = client
-              .defaultEmbed()
-              .setTitle('FAQ')
-              .setDescription(
-                faq
-                  .map(({ question }, index) => `${index + 1}. ${question}`)
-                  .join('\n')
-              );
-
-            // Create a select menu for the user to select a question, and then edit the embed to show the answer.
-            const selectMenu = [
-              new MessageActionRow().addComponents(
-                new MessageSelectMenu().setCustomId('faq').setOptions(
-                  faq.map(
-                    ({ question }, index): MessageSelectOptionData => ({
-                      label: `${index + 1}. ${question}`,
-                      value: index.toString(),
-                    })
-                  )
-                )
-              ),
-            ];
-
-            const msg = await i.editReply({
-              embeds: [embed],
-              components: selectMenu,
-            });
-
-            if (!(msg instanceof Message)) return;
-
-            const collector = msg.createMessageComponentCollector({
-              time: 10 * 60 * 1000, // 10 minutes
-            });
-
-            collector.on('collect', async (int) => {
-              if (!int.isSelectMenu()) return;
-
-              int.deferUpdate();
-
-              const faq = await getFaq();
-
-              const answer = faq[parseInt(int.values[0])];
-
-              embed.setDescription(
-                `**${answer.question}**\n` +
-                  turndownService.turndown(answer.answer)
-              );
-
-              await i.editReply({ embeds: [embed] });
-            });
-
-            collector.on('end', async () => {
-              msg.edit({ components: [] });
-            });
-
-            break;
-          }
-
-          default: {
-            await i.editReply("Not implemented or doesn't exist.");
-            break;
-          }
-        }
+      case 'ticket': {
+        await handleTicketInteraction(client, i);
         break;
       }
